Hw5 + bonus task

+++ model/src/main/java/edu/java/exceptions/BadRequestException/BadRequestException.java
@@ -4,16 +4,16 @@ import java.util.List;
 
 public abstract class BadRequestException extends Exception {
     protected static final String DEFAULT_MESSAGE = "Bad request";
-    private final List<Integer> ids;
+    private final List<Long> ids;
     private final String uri;
 
-    protected BadRequestException(List<Integer> ids, String uri, String message) {
+    protected BadRequestException(List<Long> ids, String uri, String message) {
         super(message);
         this.ids = ids;
         this.uri = uri;
     }
 
-    public List<Integer> ids() {
+    public List<Long> ids() {
         return ids;
     }
