+++ model/src/main/java/edu/java/exceptions/BadRequestException/IncorrectDataException.java
@@ -3,23 +3,23 @@ package edu.java.exceptions.BadRequestException;
 import java.util.List;
 
 public class IncorrectDataException extends BadRequestException {
-    public IncorrectDataException(List<Integer> ids, String uri, String message) {
+    public IncorrectDataException(List<Long> ids, String uri, String message) {
         super(ids, uri, message);
     }
 
-    public IncorrectDataException(List<Integer> ids, String uri) {
+    public IncorrectDataException(List<Long> ids, String uri) {
         this(ids, uri, DEFAULT_MESSAGE);
     }
 
-    public IncorrectDataException(Integer id, String uri) {
+    public IncorrectDataException(long id, String uri) {
         this(List.of(id), uri);
     }
 
-    public IncorrectDataException(Integer id) {
+    public IncorrectDataException(long id) {
         this(id, "");
     }
 
-    public Integer id() {
+    public long id() {
         return this.ids().getFirst();
     }
 }
