+++ model/src/main/java/edu/java/exceptions/NotFoundException/ChatNotRegisteredException.java
@@ -1,15 +1,15 @@
 package edu.java.exceptions.NotFoundException;
 
 public class ChatNotRegisteredException extends NotFoundException {
-    public ChatNotRegisteredException(Integer id, String uri, String message) {
+    public ChatNotRegisteredException(long id, String uri, String message) {
         super(id, uri, message);
     }
 
-    public ChatNotRegisteredException(Integer id, String uri) {
+    public ChatNotRegisteredException(long id, String uri) {
         this(id, uri, DEFAULT_MESSAGE);
     }
 
-    public ChatNotRegisteredException(Integer id) {
+    public ChatNotRegisteredException(long id) {
         this(id, "", DEFAULT_MESSAGE);
     }
 }
